+++ driver/sql/postgres/pgset/schema.go
@@ -11,14 +11,17 @@ import (
 //go:embed schema.sql
 var schema string
 
-// createSchema creates the PostgreSQL schema elements required by [BinaryStore].
-func createSchema(
-	ctx context.Context,
-	db *sql.DB,
-) error {
+// Provision creates the PostgreSQL schema and tables used by the store if they
+// do not already exist.
+//
+// The store also creates the schema on first use if it does not exist.
+// Provision allows infrastructure to be created ahead of time, for example as
+// part of a deployment pipeline, so that the application itself does not need
+// DDL permissions.
+func (s *BinaryStore) Provision(ctx context.Context) error {
 	return pgerror.Retry(
 		ctx,
-		db,
+		s.DB,
 		func(tx *sql.Tx) error {
 			_, err := tx.ExecContext(ctx, schema)
 			return err
