+++ driver/sql/postgres/pgset/store.go
@@ -21,17 +21,6 @@ func NewBinaryStore(db *sql.DB) *BinaryStore {
 	return &BinaryStore{DB: db}
 }
 
-// Provision creates the PostgreSQL schema and tables used by the store if they
-// do not already exist.
-//
-// The store also creates the schema on first use if it does not exist.
-// Provision allows infrastructure to be created ahead of time, for example as
-// part of a deployment pipeline, so that the application itself does not need
-// DDL permissions.
-func (s *BinaryStore) Provision(ctx context.Context) error {
-	return createSchema(ctx, s.DB)
-}
-
 // Open returns the set with the given name.
 func (s *BinaryStore) Open(ctx context.Context, name string) (set.BinarySet, error) {
 	id, err := s.getID(ctx, name)
@@ -66,7 +55,7 @@ func (s *BinaryStore) getID(ctx context.Context, name string) (uint64, error) {
 			return 0, fmt.Errorf("cannot scan set ID: %w", err)
 		}
 
-		if err := createSchema(ctx, s.DB); err != nil {
+		if err := s.Provision(ctx); err != nil {
 			return 0, fmt.Errorf("cannot create set schema: %w", err)
 		}
 	}
